refactor types, messages, and operators
* clean up usage of ofMessageType
* move types to separate file
* plan on moving messages and operators
